Use separate command for generating files(manifest, ytb script)
When same command is used for generating and packaging, packaged
script is empty file.
